[KKIP-88] Pages updated with data, components refactored to Main.tsx
